+++ Egyptian_Horror_Story/Egyptian_Horror_Story/GraphicsData.cpp
@@ -90,7 +90,7 @@ void GraphicsData::createVerticeArray(int key, EntityStruct::VertexStruct* verti
 	this->mNrOfVertices[key] = nrOfVertices;
 	this->mVertices[key] = new EntityStruct::VertexStruct[nrOfVertices];
 
-	for (size_t i = 0; i < nrOfVertices; i++)
+	for (int i = 0; i < nrOfVertices; i++)
 	{
 		this->mVertices.at(key)[i] = vertices[i];
 	}
